+++ src/ServiceProvider.php
@@ -3,12 +3,24 @@
 namespace Juy\Providers;
 
 use Illuminate\Foundation\AliasLoader;
-use Illuminate\Support\ServiceProvider as IlluminateServiceProvider;
+use Illuminate\Support\ServiceProvider as BaseServiceProvider;
 
-class ServiceProvider extends IlluminateServiceProvider
+/**
+ * Class ServiceProvider
+ * 
+ * @package Juy\Providers
+ */
+class ServiceProvider extends BaseServiceProvider
 {
     /**
-     * Indicates if loading of the provider is deferred.
+     * Package name
+     *
+     * @const string
+     */
+    const PACKAGE_NAME = 'providers';
+
+    /**
+     * Indicates if loading of the provider is deferred
      *
      * @var bool
      */
@@ -22,10 +34,8 @@ class ServiceProvider extends IlluminateServiceProvider
     public function register()
     {
         // Default package configuration
-        $this->mergeConfigFrom(
-            __DIR__ . '/config/providers.php', 'providers'
-        );
-
+        $this->mergeConfig();
+        
         // Count the array recursively. Empty config count is 10.
         if (count($this->app['config']->get('providers'), true) !== 10)
         {
@@ -75,17 +85,31 @@ class ServiceProvider extends IlluminateServiceProvider
     }
 
     /**
-     * Publish the config file.
+     * Default package configuration
+     *
+     * @return void
+     */
+    protected function mergeConfig()
+    {
+        $this->mergeConfigFrom(
+            $this->packagePath('config/config.php'), self::PACKAGE_NAME
+        );
+    }
+
+    /**
+     * Publish the config file
+     *
+     * @return void
      */
     protected function publishConfig()
     {
         $this->publishes([
-            __DIR__ . '/config/providers.php' => config_path('providers.php')
+            $this->packagePath('config/config.php') => config_path(self::PACKAGE_NAME . '.php')
         ], 'config');
     }
-
+    
     /**
-     * Register providers.
+     * Register providers
      *
      * @param array $providers
      * @return void
@@ -99,7 +123,7 @@ class ServiceProvider extends IlluminateServiceProvider
     }
 
     /**
-     * Register Aliases.
+     * Register Aliases
      *
      * @param array $aliases
      * @return void
@@ -112,5 +136,16 @@ class ServiceProvider extends IlluminateServiceProvider
         {
             $aliasLoader->alias($alias, $facade);
         }
+    }
+    
+    /**
+     * Loads a path relative to the package base directory
+     *
+     * @param string $path
+     * @return string
+     */
+    protected function packagePath($path = '')
+    {
+        return sprintf('%s/../%s', __DIR__, $path);
     }
 }